light restructuring of folders
