+++ src/Details/details.css
