+++ src/Details/details.js
